Move dependency to dependencyinterfaces package
